cloud-function/requirements.txt
# Python dependencies for BigQuery setup script
google-cloud-bigquery>=3.11.4